Cleaned up unused code and renamed test class
